Переименовали сущность/поля/страницы/сервис/контроллер для работы с подобъектами

+++ src/main/java/com/art/model/supporting/AfterCashing.java
@@ -7,40 +7,39 @@ import lombok.ToString;
 import javax.persistence.*;
 import java.io.Serializable;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 
 @EqualsAndHashCode
 @ToString
 @Entity
 @Table(name = "AfterCashing")
 public class AfterCashing implements Serializable {
-    private BigInteger id;
-    private BigInteger oldId;
+    private Long id;
+    private Long oldId;
     private BigDecimal oldValue;
 
     public AfterCashing() {
     }
 
-    public AfterCashing(BigInteger oldId, BigDecimal oldValue) {
+    public AfterCashing(Long oldId, BigDecimal oldValue) {
         this.oldId = oldId;
         this.oldValue = oldValue;
     }
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
     @Column(name = "OldId")
-    public BigInteger getOldId() {
+    public Long getOldId() {
         return oldId;
     }
 
-    public void setOldId(BigInteger oldId) {
+    public void setOldId(Long oldId) {
         this.oldId = oldId;
     }
